Renaming iree/tools/utils/ -> iree/tooling/. This matches iree/testing/ now that the directory no longer contains tools but just tooling-related code.
