+++ experimental/web/sample_dynamic/main.c
@@ -49,7 +49,7 @@ void unload_program(iree_program_state_t* program_state);
 //
 // * |function_name| is the fully qualified function name, like 'module.abs'.
 // * |inputs| is a semicolon delimited list of VM scalars and buffers, as
-//   described in iree/tools/utils/vm_util and used in IREE's CLI tools.
+//   described in iree/tooling/vm_util and used in IREE's CLI tools.
 //   For example, the CLI `--function_input=f32=1 --function_input=f32=2`
 //   should be passed here as `f32=1;f32=2`.
 // * |iterations| is the number of times to call the function, for benchmarking
