+++ runtime/src/iree/tooling/device_util.c
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/tools/utils/device_util.h"
+#include "iree/tooling/device_util.h"
 
 #include "iree/base/internal/call_once.h"
 #include "iree/base/internal/flags.h"
