+++ runtime/src/iree/tooling/vm_util.cc
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/tools/utils/vm_util.h"
+#include "iree/tooling/vm_util.h"
 
 #include <cerrno>
 #include <cstdint>
