+++ runtime/src/iree/tooling/vm_util.h
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef IREE_TOOLS_UTILS_VM_UTIL_H_
-#define IREE_TOOLS_UTILS_VM_UTIL_H_
+#ifndef IREE_TOOLING_VM_UTIL_H_
+#define IREE_TOOLING_VM_UTIL_H_
 
 #include <iostream>
 #include <ostream>
@@ -54,4 +54,4 @@ inline Status PrintVariantList(iree_vm_list_t* variant_list,
 
 }  // namespace iree
 
-#endif  // IREE_TOOLS_UTILS_VM_UTIL_H_
+#endif  // IREE_TOOLING_VM_UTIL_H_
