+++ tools/BUILD
@@ -33,8 +33,8 @@ cc_binary(
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/hal",
         "//runtime/src/iree/modules/hal",
-        "//runtime/src/iree/tools/utils:device_util",
-        "//runtime/src/iree/tools/utils:vm_util",
+        "//runtime/src/iree/tooling:device_util",
+        "//runtime/src/iree/tooling:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
@@ -53,9 +53,9 @@ cc_binary(
         "//runtime/src/iree/base/internal:path",
         "//runtime/src/iree/hal",
         "//runtime/src/iree/testing:benchmark",
-        "//runtime/src/iree/tools/utils:device_util",
-        "//runtime/src/iree/tools/utils:trace_replay",
-        "//runtime/src/iree/tools/utils:yaml_util",
+        "//runtime/src/iree/tooling:device_util",
+        "//runtime/src/iree/tooling:trace_replay",
+        "//runtime/src/iree/tooling:yaml_util",
         "//runtime/src/iree/vm",
         "@com_github_yaml_libyaml//:yaml",
     ],
@@ -77,8 +77,8 @@ cc_binary(
         "//runtime/src/iree/modules/check",
         "//runtime/src/iree/modules/hal",
         "//runtime/src/iree/testing:gtest",
-        "//runtime/src/iree/tools/utils:device_util",
-        "//runtime/src/iree/tools/utils:vm_util",
+        "//runtime/src/iree/tooling:device_util",
+        "//runtime/src/iree/tooling:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
     ],
@@ -139,8 +139,8 @@ cc_binary(
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/hal",
         "//runtime/src/iree/modules/hal",
-        "//runtime/src/iree/tools/utils:device_util",
-        "//runtime/src/iree/tools/utils:vm_util",
+        "//runtime/src/iree/tooling:device_util",
+        "//runtime/src/iree/tooling:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
@@ -165,8 +165,8 @@ cc_binary(
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/hal",
         "//runtime/src/iree/modules/hal",
-        "//runtime/src/iree/tools/utils:device_util",
-        "//runtime/src/iree/tools/utils:vm_util",
+        "//runtime/src/iree/tooling:device_util",
+        "//runtime/src/iree/tooling:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
@@ -182,9 +182,9 @@ cc_binary(
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/base/internal:path",
         "//runtime/src/iree/hal",
-        "//runtime/src/iree/tools/utils:device_util",
-        "//runtime/src/iree/tools/utils:trace_replay",
-        "//runtime/src/iree/tools/utils:yaml_util",
+        "//runtime/src/iree/tooling:device_util",
+        "//runtime/src/iree/tooling:trace_replay",
+        "//runtime/src/iree/tooling:yaml_util",
         "//runtime/src/iree/vm",
         "@com_github_yaml_libyaml//:yaml",
     ],
@@ -201,10 +201,10 @@ cc_binary(
         "//runtime/src/iree/base/internal:path",
         "//runtime/src/iree/hal",
         "//runtime/src/iree/modules/hal",
-        "//runtime/src/iree/tools/utils:cpu_features",
-        "//runtime/src/iree/tools/utils:device_util",
-        "//runtime/src/iree/tools/utils:trace_replay",
-        "//runtime/src/iree/tools/utils:yaml_util",
+        "//runtime/src/iree/tooling:cpu_features",
+        "//runtime/src/iree/tooling:device_util",
+        "//runtime/src/iree/tooling:trace_replay",
+        "//runtime/src/iree/tooling:yaml_util",
         "//runtime/src/iree/vm",
         "@com_github_yaml_libyaml//:yaml",
     ],
