+++ tools/iree-run-trace-main.c
@@ -12,9 +12,9 @@
 #include "iree/base/internal/flags.h"
 #include "iree/base/internal/path.h"
 #include "iree/hal/api.h"
-#include "iree/tools/utils/device_util.h"
-#include "iree/tools/utils/trace_replay.h"
-#include "iree/tools/utils/yaml_util.h"
+#include "iree/tooling/device_util.h"
+#include "iree/tooling/trace_replay.h"
+#include "iree/tooling/yaml_util.h"
 #include "iree/vm/api.h"
 
 IREE_FLAG(bool, trace_execution, false, "Traces VM execution to stderr.");
